feat(audio-reader): inline the add-card flow as a slide-over in the term card
Clicking add-card now slides an inline form over the translation from the right
instead of opening a global modal — in the footer on mobile and in the popover
on desktop. The term card's height change drives the footer's height animation;
saving dismisses the term, cancelling slides back. Removes the add-card modal and
its composable.
