+++ src/views/audio-reader/term-popover/term-card.vue
@@ -6,7 +6,10 @@ import UiButton from '@/components/ui-kit/button.vue'
 import UiDivider from '@/components/ui-kit/divider.vue'
 import UiTag from '@/components/ui-kit/tag.vue'
 import AddCardControl from './add-card-control.vue'
-import { useAddCardModal } from '@/composables/modals/use-add-card-modal'
+import AddCardPanel from './add-card-panel.vue'
+import { cardSlideEnter, cardSlideLeave } from '@/utils/animations/card-slide'
+
+type AddCardDraft = { front: string; back: string; deck_id: number | null }
 
 const {
   term,
@@ -29,11 +32,11 @@ const emit = defineEmits<{
 
 const { t } = useI18n()
 const translate = useTranslateTermMutation()
-const add_card_modal = useAddCardModal()
 
 const result = ref<TranslationResult | null>(null)
 const is_loading = ref(false)
 const error_key = ref<string | null>(null)
+const adding = ref<AddCardDraft | null>(null)
 
 async function fetchTranslation() {
   result.value = null
@@ -51,13 +54,23 @@ async function fetchTranslation() {
   }
 }
 
-// Hand the translation off to the add-card modal and dismiss the popover — the
-// modal lives in the global stack, so closing the popover here doesn't unmount it.
+// Slide the add-card panel over the translation, pre-filled with the term and
+// its translation. The panel handles the save; closing the term is left to its
+// `saved`/`cancel` outcomes.
 function onAddCard(deck_id: number | null) {
   if (!result.value) return
 
-  add_card_modal.open(term, result.value.translation, deck_id)
+  adding.value = { front: term, back: result.value.translation, deck_id }
+}
+
+// Saving dismisses the whole term (back to the toolbar on mobile); cancelling
+// slides back to the translation.
+function onPanelSaved() {
   emit('close')
+}
+
+function onPanelCancel() {
+  adding.value = null
 }
 
 // The card only mounts while a term is showing, so fetch on mount and whenever
@@ -74,157 +87,174 @@ watch(
 <template>
   <div
     data-testid="term-card"
-    class="flex flex-col [--skeleton-sheen:var(--color-brown-200)] dark:[--skeleton-sheen:var(--color-grey-400)]"
+    class="relative overflow-hidden [--skeleton-sheen:var(--color-brown-200)] dark:[--skeleton-sheen:var(--color-grey-400)]"
   >
-    <div
-      v-if="show_back"
-      data-testid="term-card__controls"
-      class="mb-8 flex items-center justify-between gap-3"
-    >
-      <ui-button
-        data-testid="term-card__back"
-        data-theme="brown-100"
-        data-theme-dark="stone-900"
-        icon-left="arrow-back"
-        icon-only
-        size="base"
-        play-on-tap
-        :sfx="{ click: 'ui.snappy_button_5' }"
-        @click="emit('back')"
-      >
-        {{ t('audio-reader.popover.back-button') }}
-      </ui-button>
-
-      <add-card-control :disabled="!result" @add="onAddCard" />
-    </div>
-
-    <header
-      data-testid="term-card__header"
-      class="flex items-start gap-3"
-      :class="show_back ? 'justify-center' : 'justify-between'"
-    >
-      <div data-testid="term-card__term-group" class="flex min-w-0 items-center gap-2">
-        <span
-          data-testid="term-card__term"
-          class="text-7xl leading-tight wrap-break-word text-brown-700 dark:text-brown-200"
-          :class="{ 'text-center': show_back }"
+    <transition :css="false" @enter="cardSlideEnter" @leave="cardSlideLeave">
+      <div v-if="!adding" key="term" data-testid="term-card__face" class="flex flex-col">
+        <div
+          v-if="show_back"
+          data-testid="term-card__controls"
+          class="mb-8 flex items-center justify-between gap-3"
         >
-          {{ term }}
-        </span>
+          <ui-button
+            data-testid="term-card__back"
+            data-theme="brown-100"
+            data-theme-dark="stone-900"
+            icon-left="arrow-back"
+            icon-only
+            size="base"
+            play-on-tap
+            :sfx="{ click: 'ui.snappy_button_5' }"
+            @click="emit('back')"
+          >
+            {{ t('audio-reader.popover.back-button') }}
+          </ui-button>
 
-        <ui-button
-          v-if="!show_back"
-          data-testid="term-card__play-word"
-          data-theme="grey-400"
-          class="shrink-0"
-          icon-left="play"
-          icon-only
-          size="sm"
-          @click="emit('play-word')"
+          <add-card-control :disabled="!result" @add="onAddCard" />
+        </div>
+
+        <header
+          data-testid="term-card__header"
+          class="flex items-start gap-3"
+          :class="show_back ? 'justify-center' : 'justify-between'"
         >
-          {{ t('audio-reader.popover.play-word-button') }}
-        </ui-button>
+          <div data-testid="term-card__term-group" class="flex min-w-0 items-center gap-2">
+            <span
+              data-testid="term-card__term"
+              class="text-7xl leading-tight wrap-break-word text-brown-700 dark:text-brown-200"
+              :class="{ 'text-center': show_back }"
+            >
+              {{ term }}
+            </span>
+
+            <ui-button
+              v-if="!show_back"
+              data-testid="term-card__play-word"
+              data-theme="grey-400"
+              class="shrink-0"
+              icon-left="play"
+              icon-only
+              size="sm"
+              @click="emit('play-word')"
+            >
+              {{ t('audio-reader.popover.play-word-button') }}
+            </ui-button>
+          </div>
+
+          <template v-if="!show_back">
+            <add-card-control v-if="result" @add="onAddCard" />
+
+            <ui-button
+              v-else
+              data-testid="term-card__close"
+              data-theme="grey-400"
+              icon-left="close"
+              icon-only
+              size="sm"
+              @click="emit('close')"
+            >
+              {{ t('audio-reader.popover.close-button') }}
+            </ui-button>
+          </template>
+        </header>
+
+        <div
+          data-testid="term-card__reading"
+          class="mt-1 flex min-h-7 flex-wrap items-center gap-2"
+          :class="{ 'justify-center': show_back }"
+        >
+          <span v-if="result?.reading" class="text-base text-brown-700 dark:text-grey-400">
+            {{ result.reading }}
+          </span>
+
+          <span
+            v-else-if="is_loading"
+            class="term-card__skeleton h-4 w-24 rounded-2 bg-brown-500 dark:bg-grey-500"
+          />
+        </div>
+
+        <ui-divider class="my-3">
+          <template #start>
+            <span class="shrink-0 text-brown-500">{{
+              t('audio-reader.popover.definition-label')
+            }}</span>
+          </template>
+
+          <template #end>
+            <ui-tag
+              v-if="result?.pos"
+              data-theme="green-400"
+              class="shrink-0 bgx-diagonal-stripes"
+              >{{ result.pos }}</ui-tag
+            >
+          </template>
+        </ui-divider>
+
+        <div data-testid="term-card__body" class="flex min-h-18 flex-col">
+          <div
+            v-if="is_loading"
+            data-testid="term-card__loading"
+            class="flex flex-col gap-2"
+            aria-busy="true"
+            :aria-label="t('audio-reader.popover.loading')"
+          >
+            <span class="term-card__skeleton h-6 w-3/5 rounded-2 bg-brown-500 dark:bg-grey-500" />
+            <span class="term-card__skeleton h-4 w-full rounded-2 bg-brown-500 dark:bg-grey-500" />
+            <span class="term-card__skeleton h-4 w-4/5 rounded-2 bg-brown-500 dark:bg-grey-500" />
+          </div>
+
+          <p
+            v-else-if="error_key"
+            data-testid="term-card__error"
+            class="text-red-500 dark:text-red-400"
+          >
+            {{ t(error_key) }}
+          </p>
+
+          <div v-else-if="result" data-testid="term-card__result" class="flex flex-col gap-1">
+            <p
+              data-testid="term-card__translation"
+              class="text-3xl text-brown-700 capitalize dark:text-brown-200"
+            >
+              {{ result.translation }}
+            </p>
+            <p
+              v-if="result.description"
+              data-testid="term-card__description"
+              class="text-base text-brown-700 dark:text-grey-300"
+            >
+              {{ result.description }}
+            </p>
+          </div>
+        </div>
+
+        <footer data-testid="term-card__footer" class="mt-4 flex">
+          <ui-button
+            data-testid="term-card__play-from-here"
+            data-theme="brown-100"
+            data-theme-dark="stone-900"
+            icon-left="play"
+            size="xl"
+            full-width
+            play-on-tap
+            :sfx="{ click: 'ui.snappy_button_3' }"
+            @click="emit('play-from-here')"
+          >
+            {{ t('audio-reader.popover.play-from-here-button') }}
+          </ui-button>
+        </footer>
       </div>
 
-      <template v-if="!show_back">
-        <add-card-control v-if="result" @add="onAddCard" />
-
-        <ui-button
-          v-else
-          data-testid="term-card__close"
-          data-theme="grey-400"
-          icon-left="close"
-          icon-only
-          size="sm"
-          @click="emit('close')"
-        >
-          {{ t('audio-reader.popover.close-button') }}
-        </ui-button>
-      </template>
-    </header>
-
-    <div
-      data-testid="term-card__reading"
-      class="mt-1 flex min-h-7 flex-wrap items-center gap-2"
-      :class="{ 'justify-center': show_back }"
-    >
-      <span v-if="result?.reading" class="text-base text-brown-700 dark:text-grey-400">
-        {{ result.reading }}
-      </span>
-
-      <span
-        v-else-if="is_loading"
-        class="term-card__skeleton h-4 w-24 rounded-2 bg-brown-500 dark:bg-grey-500"
+      <add-card-panel
+        v-else
+        key="add"
+        :front="adding.front"
+        :back="adding.back"
+        :deck_id="adding.deck_id"
+        @cancel="onPanelCancel"
+        @saved="onPanelSaved"
       />
-    </div>
-
-    <ui-divider class="my-3">
-      <template #start>
-        <span class="shrink-0 text-brown-500">{{
-          t('audio-reader.popover.definition-label')
-        }}</span>
-      </template>
-
-      <template #end>
-        <ui-tag v-if="result?.pos" data-theme="green-400" class="shrink-0 bgx-diagonal-stripes">{{
-          result.pos
-        }}</ui-tag>
-      </template>
-    </ui-divider>
-
-    <div data-testid="term-card__body" class="flex min-h-18 flex-col">
-      <div
-        v-if="is_loading"
-        data-testid="term-card__loading"
-        class="flex flex-col gap-2"
-        aria-busy="true"
-        :aria-label="t('audio-reader.popover.loading')"
-      >
-        <span class="term-card__skeleton h-6 w-3/5 rounded-2 bg-brown-500 dark:bg-grey-500" />
-        <span class="term-card__skeleton h-4 w-full rounded-2 bg-brown-500 dark:bg-grey-500" />
-        <span class="term-card__skeleton h-4 w-4/5 rounded-2 bg-brown-500 dark:bg-grey-500" />
-      </div>
-
-      <p
-        v-else-if="error_key"
-        data-testid="term-card__error"
-        class="text-red-500 dark:text-red-400"
-      >
-        {{ t(error_key) }}
-      </p>
-
-      <div v-else-if="result" data-testid="term-card__result" class="flex flex-col gap-1">
-        <p
-          data-testid="term-card__translation"
-          class="text-3xl text-brown-700 capitalize dark:text-brown-200"
-        >
-          {{ result.translation }}
-        </p>
-        <p
-          v-if="result.description"
-          data-testid="term-card__description"
-          class="text-base text-brown-700 dark:text-grey-300"
-        >
-          {{ result.description }}
-        </p>
-      </div>
-    </div>
-
-    <footer data-testid="term-card__footer" class="mt-4 flex">
-      <ui-button
-        data-testid="term-card__play-from-here"
-        data-theme="brown-100"
-        data-theme-dark="stone-900"
-        icon-left="play"
-        size="xl"
-        full-width
-        play-on-tap
-        :sfx="{ click: 'ui.snappy_button_3' }"
-        @click="emit('play-from-here')"
-      >
-        {{ t('audio-reader.popover.play-from-here-button') }}
-      </ui-button>
-    </footer>
+    </transition>
   </div>
 </template>
